Long-Float fix

+++ mlc/model/cnn/model.py
@@ -51,7 +51,7 @@ class CNN(BaseModel):
 
     def evaluate_loss(self, Y_pred, Y):
         # F.cross_entropy expects logits, not probabilities
-        Y_pred = Y_pred.view(-1)
+        Y_pred = Y_pred.float().view(-1)
         return F.binary_cross_entropy(Y_pred, Y)
 
     def forward(self, x):
